civic_rag/requirements.txt
# Python dependencies for Civic RAG
streamlit
langchain>=0.1.0
langchain-community
langchain-groq
chromadb>=0.4.0
pypdf>=3.0.0
tiktoken
python-dotenv
tabulate
sqlalchemy
PyPDF2
typing-extensions
sentence-transformers
